temp fix

+++ bot.py
@@ -473,7 +473,10 @@ async def create_ticket(guild, member):
         ticket_message = await channel.send(f"Hello {member.mention}, please describe your issue in as much detail as possible.", embed=embed)
         await ticket_message.add_reaction("🔒")
         await ticket_message.pin(reason=f'Pinned first message in #{channel.name}')
-        await channel.last_message.delete()
+        # this should be done differently - use the on_message event
+        lm = await channel.last_message
+        if lm:
+            await channel.last_message.delete()
         cursor = db.cursor()
         command = f"""INSERT INTO tickets (ticketchannel, owner, parentguild)
                         VALUES({channel.id}, {member.id}, {guild.id});"""
